speedtest: Add caf command support
This commit adds support for three new caf related commands:

caf_latency
caf_download
caf_upload

These commands reuse the same latency and transfer code as the normal
speedtests, but reach out to new urls for settings and telemetry.

Additionally, this commit adds support for an optional 'test_id'
setting.  If it is received in the test settings, it is added
to the results json that is pushed back to the cloud.

MIN-8723 MIN-8722