util_stime: Remove XSTR() and STR() defines
These came in with a commit intended to fix compile issues with
2.31 and gcc 9.x, but I don't see how they are needed for that
combination.  Since moving util_stime.h to util_common.h there are
build errors with platforms that define these elsewhere.  Lets remove
these from util_stime.h and see if anything breaks.  If a platform
breaks because of the lack of these defines, we should include them
in a platform specific header now that util_stime.h is common.

SW-838