rules: Add lede_generic.mk
This allows an openwrt unum feed makefile to call the top level
makefile in unum-sdk.  This means that the unum build process for
openwrt will be more in line with other non-openwrt based products
and that the logic in the top level makefile won't have to be
copied to the feed makefile.

SW-815