zoom_7020_openwifi: Add more onboarding scripts
This follows the onboarding methodology of other lede_generic builds
like the gl_b1300, where the device comes up by default with an
onboarding station interface looking for a minim-$(mac) ssid with a
key that matches the units serial number.  When a request is made to
onboard an extender, that network will be created on the root.

SW-815