zoom_7020_openwifi: Don't clobber wireless settings on upgrade
uci-defaults get run from factory and after a firmware upgrade.
If the device has been configured via the cloud, then we don't want
to set the default wireless settings after a firmware upgrade.
Check for the presence of the unum opmode to determine if the device
has been configured already.