minim_os: Scan for onboarding network before enableing sta0
Create the onboarding station, but keep it disabled initially.
Scan the frequencies on the 5ghz radio until the target onboarding
network is detected.  Then enable the sta0 interface to attach to it.
This can help pre-onboarding 5ghz stability on devices that don't work
well with an AP interface and unattached station interface on the same
radio.

SW-2569