telemetry_ubus: Initialize last_smbt_device_state to non-zero
If there is no USB drive present at boot, we want to send up an
empty smb_devices array.  Initialize last_smbt_device_state to
non-zero so the first time we check for a difference between the
current and last data, we see a difference and send up the empty
array.

SW-3072